+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Breaking change: Remove `k-Alert` SCSS mixin.
+- Breaking change: Remove `k-Popover` SCSS mixin.
+- Breaking change: Remove `Popover` and `CallToActionPopover`.
 - Breaking change: Remove `k-ExternalRichLink` SCSS mixin.
 - Breaking change: Remove `ExternalRichLink` component.